+++ src/terminal_sync/__main__.py
@@ -1,14 +1,6 @@
-"""Executes the terminal_sync server when the module is run"""
-
-# Standard Libraries
-from argparse import ArgumentParser
+"""Executes the terminal_sync CLI when the module is run"""
 
 # Internal Libraries
-from terminal_sync.api import run
+from terminal_sync.cli import main
 
-parser = ArgumentParser()
-parser.add_argument("--host", dest="host", default="127.0.0.1", help="The host address where the server will bind")
-parser.add_argument("--port", dest="port", default=8000, type=int, help="The host port where the server will bind")
-args = parser.parse_args()
-
-run(host=args.host, port=args.port)
+main()